+++ src/bullet/index.tsx
@@ -24,9 +24,18 @@ const StyledBullet = styled(Bullet)`
   &.kirk-bullet--map-active,
   &.kirk-bullet--map-inactive,
   &.kirk-bullet--search {
+    border-width: 3px;
+  }
+
+  &.kirk-bullet--map-active,
+  &.kirk-bullet--map-inactive {
     width: ${componentSizes.bulletSizeMap};
     height: ${componentSizes.bulletSizeMap};
-    border-width: 3px;
+  }
+
+  &.kirk-bullet--search {
+    width: ${componentSizes.bulletSizeSearch};
+    height: ${componentSizes.bulletSizeSearch};
   }
 
   &.kirk-bullet--map-active {