+++ src/datePicker/index.tsx
@@ -90,12 +90,10 @@ const StyledDatePicker = styled(DatePicker)`
     color: ${color.fadedText};
     cursor: default;
   }
-  & .DayPicker-Day--selected span,
-  & .DayPicker-Day:focus > span {
+  & .DayPicker-Day--selected span {
     background-color: ${color.primary};
     color: ${color.white};
   }
-
   & .DayPicker-Day span {
     display: inline-flex;
     justify-content: center;